Move configurator and local conf to utils and fix import
